backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
groq
chromadb
pypdf
python-dotenv
python-docx
python-pptx
openpyxl
pillow
pytesseract
